backend-django/requirements.txt
django
djangorestframework
django-cors-headers
pytz